modified:   src/main/java/com/springbootconcepts/modifyrequest/controller/UserController.java 	renamed:    src/main/java/com/springbootconcepts/modifyrequest/doain/User.java -> src/main/java/com/springbootconcepts/modifyrequest/domain/User.java
